requirements.txt
discord.py[voice]
alkana.py
beautifulsoup4
emoji
html5lib
lxml
pandas
pillow
pydub
python-dotenv
unidic-lite
gspread
oauth2client